scripts/producer/requirements.txt
kafka-python
lasio
